backend/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-limiter==3.7.0
python-dotenv==1.0.1
supabase==2.4.6
requests==2.31.0
gunicorn==22.0.0
pydantic==2.7.1
